role dropdown list

+++ CantinaUnitBV/Pages/Users/Delete.cshtml.cs
@@ -26,7 +26,7 @@ namespace CantinaUnitBV.Pages
         public User User { get; set; }
         public string ErrorMessage { get; set; }
 
-        public async Task<IActionResult> OnGetAsync(int? id, bool? saveChangesError = false)
+        public async Task<IActionResult> OnGetAsync(long? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -50,7 +50,7 @@ namespace CantinaUnitBV.Pages
             return Page();
         }
 
-        public async Task<IActionResult> OnPostAsync(int? id)
+        public async Task<IActionResult> OnPostAsync(long? id)
         {
             if (id == null)
             {
